+++ assets/javascripts/kitten/components/icons/checked-circle-icon.js
@@ -21,5 +21,5 @@ export const CheckedCircleIcon = ({
 CheckedCircleIcon.defaultProps = {
   circleColor: '#fff',
   checkedColor: '#333',
-  title: '',
+  title: 'Checked circle',
 }