+++ CHANGELOG.md
@@ -4,7 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: Add `avatarBadge` props on `CommentAvatar`.
+- Feature: Add `title` prop on `CheckedCircleIcon`.
+- Feature: Add `avatarBadge` prop on `CommentAvatar`, `Comment` and `CommentForm`.
 
 ## [2.15.0] - 2019-07-16